+++ tests/e2e/mwp-admin.get-started.e2e.ts
@@ -8,6 +8,7 @@
 
 import { expect, browser, $ } from '@wdio/globals';
 import fetch from 'node-fetch';
+import * as semver from 'semver';
 import MwpGetStartedPage from './pageobjects/mwp-admin/get-started.page.js';
 import Website from './website.js';
 import OverviewPage from './pageobjects/matomo-reporting/visitors/overview.page.js';
@@ -46,5 +47,21 @@ describe('MWP Admin > Get Started', () => {
   it('should hide the page when both steps are completed', async () => {
     await MwpGetStartedPage.enableTracking();
     expect(await $('.matomo-dashboard-container').isExisting()).toBeTruthy();
+  });
+
+  it('should display the minimum requirements notice', async () => {
+    await MwpGetStartedPage.open();
+
+    const notice = await $('#matomo-minimumrequirements');
+    if (semver.gte(`${process.env.PHP_VERSION}.0`, '8.1.0')) {
+      expect(await notice.isExisting()).toBeFalsy();
+    } else {
+      await notice.waitForDisplayed({ timeout: 30000 });
+
+      await expect(notice).toHaveText(
+        'Matomo Analytics version 6 and later will require a newer server environment',
+        { containing: true }
+      );
+    }
   });
 });